added delete monthly rating, configured some messages, added pie chart

+++ src/app/users/users.component.ts
@@ -1,9 +1,11 @@
+import { VacationLeave } from './../vacationleaves/vacationleave.model';
 import { Component, OnInit } from '@angular/core';
 import { MatTableDataSource } from '@angular/material';
 import { Router } from '@angular/router';
 import { User } from './user.model';
 import { ToastrService } from 'ngx-toastr';
 import { RepositoryFunctionService } from '../shared/repository.function-service';
+import { MonthlyRating } from '../monthlyratings/monthlyrating.model';
 
 @Component({
   selector: 'app-users',
@@ -37,10 +39,29 @@ export class UsersComponent implements OnInit {
   }
 
   public deleteUser = (id: string) => {
-    this.functionRepoService.deleteUser(id).subscribe(res => {
-      this.getUsers();
-      this.toastr.success("Uporabnik je bil izbrisan");
-    })
+    //first check if user has related data
+    this.functionRepoService.getVacationLeaves().subscribe(res => {
+      let vacationLeaves = res as VacationLeave[];
+      let userHasVacationLeave = vacationLeaves.find(x => x.UserId == id)
+      if(userHasVacationLeave === undefined){
+        this.functionRepoService.getMonthlyRatings(id).subscribe(res => {
+          let grades = res as MonthlyRating[];
+          if(grades.length === 0){
+            this.functionRepoService.deleteUser(id).subscribe(res => {
+              this.getUsers();
+              this.toastr.success("Uporabnik je bil izbrisan", '', { positionClass: 'toast-top-center' });
+            });
+          }else{
+            this.toastr.error("Uporabnik ima povezane podatke, zato odstranitev ni možna.", 
+            '', { positionClass: 'toast-top-center' });
+          }
+        });       
+      }else{
+        this.toastr.error("Uporabnik ima povezane podatke, zato odstranitev ni možna.", 
+        '', { positionClass: 'toast-top-center' });
+      }
+    });
+    
   }
 
 }
